refactor: useNumbers
