+++ app/layout.tsx
@@ -3,7 +3,6 @@ import { Providers } from './providers'
 import { Navbar } from '@/lib/shared/components/navs/Navbar'
 import { Footer } from '@/lib/shared/components/navs/Footer'
 import { getProjectConfig } from '@/lib/config/getProjectConfig'
-import '@/lib/shared/hooks/useNumbers'
 import Noise from './noise'
 import { satoshiFont } from '@/lib/assets/fonts/satoshi/satoshi'
 import NextTopLoader from 'nextjs-toploader'
