+++ lib/modules/pool/PoolList/PoolListTokensTag.tsx
@@ -4,8 +4,8 @@ import { isGyro, isLBP, isMetaStable, isPhantomStable, isStable, isWeighted } fr
 import { Box, HStack, Tag, Text } from '@chakra-ui/react'
 import { TokenIcon } from '../../tokens/TokenIcon'
 import { PoolListItem } from '../pool.types'
-import { toPercentageFormatted } from '@/lib/shared/utils/numbers'
 import { GqlChain, GqlPoolTokenDisplay } from '@/lib/shared/services/api/generated/graphql'
+import { weightFormat } from '@/lib/shared/utils/numbers'
 
 interface Props {
   pool: PoolListItem
@@ -60,7 +60,7 @@ export function PoolListTokensTag({ pool }: Props) {
                     <PoolTokenIcon token={token} chain={pool.chain} />
                   )}
                   <Text>{token.nestedTokens ? token.name : token.symbol}</Text>
-                  <Text>{toPercentageFormatted(token.weight || '')}</Text>
+                  <Text>{weightFormat(token.weight || '')}</Text>
                 </HStack>
               </Tag>
             )
