+++ test/vitest/setup-vitest.tsx
@@ -1,4 +1,3 @@
-import '@/lib/shared/hooks/useNumbers'
 import '@testing-library/jest-dom'
 import { configure } from '@testing-library/react'
 import fetch from 'cross-fetch'
